# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Fatourti Mobile Application APIS",
        "description": "API Description",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/factures": {
            "post": {
                "tags": [
                    "Factures(Invoices,Puschase Orders,Estimates)"
                ],
                "summary": "Create a new facture",
                "description": "Creates a new facture record in the database. (0:Invoices,1:Estimates,2:Puschase Orders) ",
                "operationId": "storeFacture",
                "requestBody": {
                    "description": "Facture details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "date",
                                    "due_date",
                                    "client_id",
                                    "type"
                                ],
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-20"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-07-20"
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "10.5"
                                    },
                                    "client_id": {
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "taxe_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "type": {
                                        "type": "integer",
                                        "enum": [
                                            0,
                                            1,
                                            2
                                        ],
                                        "example": "0"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Facture created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "date": "2024-06-20",
                                    "due_date": "2024-07-20",
                                    "client_id": 1
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/invoices": {
            "get": {
                "tags": [
                    "Factures(Invoices,Puschase Orders,Estimates)"
                ],
                "summary": "Get all invoices for the authenticated user",
                "description": "Retrieves all invoices associated with the authenticated user's organization.",
                "operationId": "getInvoicesForUser",
                "responses": {
                    "200": {
                        "description": "List of invoices",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "id": 1,
                                            "date": "2024-06-20",
                                            "client_id": 1
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/estimates": {
            "get": {
                "tags": [
                    "Factures(Invoices,Puschase Orders,Estimates)"
                ],
                "summary": "Get all estimates for the authenticated user",
                "description": "Retrieves all estimates associated with the authenticated user's organization.",
                "operationId": "getEstimatesForUser",
                "responses": {
                    "200": {
                        "description": "List of estimates",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "id": 2,
                                            "date": "2024-06-21",
                                            "client_id": 1
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/purchases": {
            "get": {
                "tags": [
                    "Factures(Invoices,Puschase Orders,Estimates)"
                ],
                "summary": "Get all purchase orders for the authenticated user",
                "description": "Retrieves all purchase orders associated with the authenticated user's organization.",
                "operationId": "getPurchaseOrdersForUser",
                "responses": {
                    "200": {
                        "description": "List of purchase orders",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "type": "object",
                                        "example": {
                                            "id": 3,
                                            "date": "2024-06-22",
                                            "client_id": 1
                                        }
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/factures/{id}": {
            "get": {
                "tags": [
                    "Factures(Invoices,Puschase Orders,Estimates)"
                ],
                "summary": "Get a specific facture by ID",
                "description": "Retrieves a specific facture and its associated products.",
                "operationId": "showFacture",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the facture",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facture details",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "date": "2024-06-20",
                                    "due_date": "2024-07-20",
                                    "client_id": 1
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facture not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Factures(Invoices,Puschase Orders,Estimates)"
                ],
                "summary": "Update a specific facture by ID",
                "description": "Updates an existing facture record in the database.",
                "operationId": "updateFacture",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the facture",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "description": "Facture details",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-06-21"
                                    },
                                    "due_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2024-07-21"
                                    },
                                    "discount": {
                                        "type": "number",
                                        "format": "float",
                                        "example": "15.75"
                                    },
                                    "client_id": {
                                        "type": "integer",
                                        "example": "2"
                                    },
                                    "taxe_id": {
                                        "type": "integer",
                                        "example": "3"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Facture updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                },
                                "example": {
                                    "id": 1,
                                    "date": "2024-06-21",
                                    "due_date": "2024-07-21",
                                    "client_id": 1
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Facture not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Factures(Invoices,Puschase Orders,Estimates)"
                ],
                "summary": "Delete a specific facture by ID",
                "description": "Deletes a specific facture record from the database.",
                "operationId": "deleteFacture",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the facture",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Facture deleted successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "404": {
                        "description": "Facture not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    }
}